backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Bottleneck==1.3.7
branca==0.6.0
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.11.0
Django==5.2.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
fiona==1.10.1
folium==0.15.1
fonttools==4.52.1
GDAL==3.8.4
geopandas==0.14.4
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
mapclassify==2.6.1
MarkupSafe==2.1.5
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft==1.3.8
mkl_random==1.2.5
networkx==3.3
numexpr==2.9.0
numpy==1.26.4
packaging==24.0
pandas==2.2.3
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.9.0
pyogrio==0.7.2
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.6.1
scipy==1.15.3
shapely==2.0.4
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
unicodedata2==15.1.0
urllib3==2.2.1
win-inet-pton==1.1.0
xyzservices==2023.10.1